python_backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.2.0
transformers==4.33.2
torch==2.0.1
sentence-transformers==2.2.2
spacy==3.6.1
